% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmd.R
\name{kmmd}
\alias{kmmd}
\title{Compute the kmmd for one sample \code{y}}
\usage{
kmmd(y, sigma = 1, threshold = Inf)
}
\arguments{
\item{y}{numeric Vector}

\item{sigma}{numeric kernel standard deviation}

\item{threshold}{numeric filter out values for exponentiation.}
}
\description{
Compute the kmmd for one sample \code{y}
}
