% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmd.R
\name{MMD}
\alias{MMD}
\title{Compute the MMD between two samples}
\usage{
MMD(y, x, y_kmmd = NULL, sigma = 1, bias = FALSE, threshold = Inf,
  approx_exp = 0)
}
\arguments{
\item{y}{numeric vector.}

\item{x}{numeric vector.}

\item{y_kmmd}{precomputed first term in MMD calculation.}

\item{sigma}{numeric kernel standard deviation.}

\item{bias}{logical; if \code{TRUE} the biased MMD is computed rather than the unbiased MMD. This can be useful since the biased MMD is always positive.}

\item{threshold}{numeric filter out values for exponentiation.}

\item{approx_exp}{integer; if 0 the usual function for the exponential distribution is used; if 1 a much faster but less accurate version of the exponential distribution is used.}
}
\description{
This function returns the estimator for the two-sample MMD.
}
\examples{
set.seed(1)
y <- rnorm(2000)
x <- rnorm(2000, 5)

MMD_1 <- MMD(y, x)
MMD_1

# Precompute y_kmmd for faster speed
y_kmmd <- kmmd(y)

MMD_2 <- MMD(y, x, y_kmmd)
MMD_2


system.time(MMD_1 <- MMD(y, x))
system.time(MMD_2 <- MMD(y, x, y_kmmd))

# Different sigma

MMD_4 <- MMD(y, x, sigma = 0.5)
}
\references{
Gretton, Arthur, et al. "A kernel method for the two-sample-problem." Advances in neural information processing systems. 2007.
}
