% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmd_l.R
\name{MMD_l}
\alias{MMD_l}
\title{Compute the MMD between two samples in linear time}
\usage{
MMD_l(y, x, var = 1, permute = FALSE)
}
\arguments{
\item{y}{numeric vector, method for matrix input not yet implemented which means observations must be 1D.}

\item{x}{numeric vector, method for matrix input not yet implemented which means observations must be 1D.}

\item{var}{numeric kernel variance.}

\item{permute}{logical Should \code{y} and \code{x} be permuted before MMD_l computation?}
}
\description{
Compute the MMD between two samples in linear time
}
\examples{
y <- rnorm(2000)
x <- rnorm(2000, 5)

MMD_full <- MMD(y, x)
MMD_linear <- MMD_l(y, x)

MMD_full
MMD_linear

}
\references{
Gretton, Arthur, et al. "A kernel two-sample test." Journal of Machine Learning Research 13.Mar (2012): 723-773.
}
