% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmd_l_multi.R
\name{MMD_l_multi}
\alias{MMD_l_multi}
\title{Compute the permuted Jack-knife MMD_linear}
\usage{
MMD_l_multi(x_obs, x_sim, k, sigma = 1)
}
\arguments{
\item{x_obs}{Numeric Vector}

\item{x_sim}{Numeric Vector}

\item{sigma}{Numeric Kernel size}
}
\description{
Compute the permuted Jack-knife MMD_linear
}
