% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmd.R
\name{MMD}
\alias{MMD}
\title{Compute the MMD between two samples}
\usage{
MMD(x_obs, x_sim, x_obs_kmmd = NULL, a = NULL, rbf = kernlab::rbfdot(sigma
  = 0.5))
}
\arguments{
\item{x_obs}{Numeric Vector}

\item{x_sim}{Numeric Vector}

\item{x_obs_kmmd}{Precomputed first term in MMD calc}

\item{a}{Termy term for the term}

\item{rbf}{The radial basis function}
}
\description{
Compute the MMD between two samples
}
