% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmd.R
\name{MMD}
\alias{MMD}
\title{Compute the MMD between two samples}
\usage{
MMD(y, x, y_kmmd = NULL, var = 1, bias = FALSE, threshold = Inf,
  approx_exp = 0, sigma = NULL, w_y = NULL, w_x = NULL)
}
\arguments{
\item{y}{either a numeric vector or matrix with number of rows equal to number of observations and number of columns equal to dimension of observations.}

\item{x}{either a numeric vector or matrix with number of rows equal to number of observations and number of columns equal to dimension of observations.}

\item{y_kmmd}{precomputed first term in MMD calculation.}

\item{var}{matrix kernel variance covariance matrix.}

\item{bias}{logical; if \code{TRUE} the biased MMD is computed rather than the unbiased MMD. This can be useful since the biased MMD is always positive.}

\item{threshold}{numeric filter out values for exponentiation.}

\item{approx_exp}{integer; if 0 the usual function for the exponential distribution is used; if 1 a much faster but less accurate version of the exponential distribution is used.}

\item{sigma}{numeric DEPRECATED square root of variance.}

\item{w_y}{numeric weights for y}

\item{w_x}{numeric weights for x}
}
\description{
This function returns the estimator for the two-sample MMD.
}
\examples{
set.seed(1)
y <- rnorm(2000)
x <- rnorm(2000, 5)

MMD_1 <- MMD(y, x)
MMD_1

# Precompute y_kmmd for faster speed
y_kmmd <- kmmd(y)

MMD_2 <- MMD(y, x, y_kmmd)
MMD_2


system.time(MMD_1 <- MMD(y, x))
system.time(MMD_2 <- MMD(y, x, y_kmmd))

# Different var

MMD_4 <- MMD(y, x, var = 0.25)
}
\references{
Gretton, Arthur, et al. "A kernel method for the two-sample-problem." Advances in neural information processing systems. 2007.
}
