% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmd.R
\name{MMD}
\alias{MMD}
\title{Compute the MMD between two samples}
\usage{
MMD(y, x, y_kmmd = NULL, sigma = 1, bias = FALSE, threshold = Inf,
  approx_exp = 0)
}
\arguments{
\item{y}{Numeric Vector}

\item{x}{Numeric Vector}

\item{y_kmmd}{Precomputed first term in MMD calculation.}

\item{sigma}{Numeric Kernel size}

\item{bias}{Logical Should the biased or unbiased MMD be computed?}

\item{threshold}{Filter out values for exponentiation}
}
\description{
This function returns the estimator for the two-sample MMD.
}
\examples{
y <- rnorm(2000)
x <- rnorm(2000, 5)

MMD_1 <- MMD(y, x)
MMD_1

# Precompute y_kmmd for faster speed
y_kmmd <- kmmd(y)

MMD_2 <- MMD(y, x, y_kmmd)
MMD_2


system.time(MMD_1 <- MMD(y, x))
system.time(MMD_2 <- MMD(y, x, y_kmmd))

# Different sigma

MMD_4 <- MMD(y, x, sigma = 0.5)
}
\references{
Gretton, Arthur, et al. "A kernel method for the two-sample-problem." Advances in neural information processing systems. 2007.
}
