% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmd_l.R
\name{MMD_l}
\alias{MMD_l}
\title{Compute the MMD between two samples in linear time}
\usage{
MMD_l(x_obs, x_sim, sigma = 1)
}
\arguments{
\item{x_obs}{Numeric Vector}

\item{x_sim}{Numeric Vector}

\item{sigma}{Numeric Kernel size}
}
\description{
Compute the MMD between two samples in linear time
}
\examples{
x_obs <- rnorm(2000)
x_sim <- rnorm(2000, 5)

MMD_full <- MMD(x_obs, x_sim)
MMD_linear <- MMD_l(x_obs, x_sim)

MMD_full
MMD_linear

}
